update for torch1

+++ pytorch/torchsparseattn/sparsemax.py
@@ -17,7 +17,7 @@ from .base import _BaseBatchProjection
 def project_simplex(v, z=1):
     v_sorted, _ = torch.sort(v, dim=0, descending=True)
     cssv = torch.cumsum(v_sorted, dim=0) - z
-    ind = torch.arange(1, 1 + len(v))
+    ind = torch.arange(1, 1 + len(v)).to(dtype=v.dtype)
     cond = v_sorted - cssv / ind > 0
     rho = ind.masked_select(cond)[-1]
     tau = cssv.masked_select(cond)[-1] / rho
